Update Finnish translation
The `spaceAnd()` and `spaceAndSpace()` functions are missing the spaces. I don't really know Finnish but when using your library in a personal project I run into this and checked with Google translate. 
I also compared it with the same functions in other languages and the spaces are indeed missing here.